Separate peak finding capabilities into a separate PeakFinder class
