+++ becquerel/core/peakfinder.py
@@ -1,0 +1,307 @@
+# -*- coding: utf-8 -*-
+
+"""Spectral peak search using convolutions."""
+
+from __future__ import print_function
+import matplotlib.pyplot as plt
+import numpy as np
+from .spectrum import Spectrum
+
+
+class PeakFilter(object):
+    """An energy-dependent kernel that can be convolved with a spectrum.
+
+    To detect lines, a kernel should have a positive component in the center
+    and negative wings to subtract the continuum, e.g., a boxcar:
+
+    +2|     ┌───┐     |
+      |     │   │     |
+     0|─┬───┼───┼───┬─|
+    -1| └───┘   └───┘ |
+
+    The positive part is meant to detect the peak, and the negative part to
+    sample the continuum under the peak.
+
+    The kernel should sum to 0.
+
+    The width of the kernel scales proportionally to the square root
+    of the channels, with a minimum value set by 'fwhm_at_0'.
+
+    """
+
+    def __init__(self, ref_channel, ref_fwhm, fwhm_at_0=1.):
+        """Initialize with a reference line position and FWHM in channels."""
+        assert ref_channel > 0
+        assert ref_fwhm > 0
+        self.ref_channel = float(ref_channel)
+        self.ref_fwhm = float(ref_fwhm)
+        self.fwhm_at_0 = float(fwhm_at_0)
+
+    def fwhm(self, channel):
+        """Calculate the expected FWHM at the given channel."""
+        # f(x)^2 = f0^2 + k x^2
+        # f1^2 = f0^2 + k x1^2
+        # k = (f1^2 - f0^2) / x1^2
+        # f(x)^2 = f0^2 + (f1^2 - f0^2) (x/x1)^2
+        f0 = self.fwhm_at_0
+        f1 = self.ref_fwhm
+        x1 = self.ref_channel
+        fwhm_sqr = f0**2 + (f1**2 - f0**2) * (channel / x1)**2
+        return np.sqrt(fwhm_sqr)
+
+    def kernel(self, channel, n_channels):
+        """Generate the kernel for the given channel."""
+        raise NotImplementedError
+
+    def kernel_matrix(self, n_channels):
+        """Build a matrix of the kernel evaluated at each channel."""
+        kern = np.zeros((n_channels, n_channels))
+        for j in range(n_channels):
+            kern[:, j] = self.kernel(j, n_channels)
+        kern_pos = +1 * kern.clip(0, np.inf)
+        kern_neg = -1 * kern.clip(-np.inf, 0)
+        # normalize negative part to be equal to the positive part
+        kern_neg *= kern_pos.sum(axis=0) / kern_neg.sum(axis=0)
+        return kern_pos - kern_neg
+
+    def plot_matrix(self, n_channels):
+        """Plot the matrix of kernels evaluated across the channels."""
+        kern_mat = self.kernel_matrix(n_channels)
+        kern_min = kern_mat.min()
+        kern_max = kern_mat.max()
+        kern_min = min(kern_min, -1 * kern_max)
+        kern_max = max(kern_max, -1 * kern_min)
+
+        plt.imshow(
+            kern_mat.T[::-1, :], cmap=plt.get_cmap('bwr'),
+            vmin=kern_min, vmax=kern_max,
+            extent=[n_channels, 0, 0, n_channels])
+        plt.colorbar()
+        plt.xlabel('Input channel')
+        plt.ylabel('Output channel')
+        plt.gca().set_aspect('equal')
+
+    def convolve(self, data):
+        """Convolve this kernel with the data."""
+        kern_mat = self.kernel_matrix(len(data))
+        kern_mat_pos = +1 * kern_mat.clip(0, np.inf)
+        kern_mat_neg = -1 * kern_mat.clip(-np.inf, 0)
+        peak_plus_bkg = np.dot(kern_mat_pos, data)
+        bkg = np.dot(kern_mat_neg, data)
+        signal = np.dot(kern_mat, data)
+        noise = np.sqrt(np.dot(kern_mat**2, data))
+        snr = np.zeros_like(signal)
+        snr[noise > 0] = signal[noise > 0] / noise[noise > 0]
+        return peak_plus_bkg, bkg, signal, noise, snr
+
+
+class BoxcarPeakFilter(PeakFilter):
+    """A spectral kernel that is a boxcar with negative wings.
+
+    The kernel is proportional to this form:
+
+    +2|     ┌───┐     |
+      |     │   │     |
+     0|─┬───┼───┼───┬─|
+    -1| └───┘   └───┘ |
+
+    """
+
+    def kernel(self, channel, n_channels):
+        """Generate the kernel for the given channel."""
+        n_center = int(np.ceil(self.fwhm(channel)))
+        if n_center % 2 == 0:
+            n_center += 1
+        kernel0 = -0.5 * np.ones(n_center)
+        kernel0 = np.append(kernel0, np.ones(n_center))
+        kernel0 = np.append(kernel0, -0.5 * np.ones(n_center))
+        n_side = len(kernel0) // 2
+        kernel = np.zeros(2 * n_side + n_channels)
+        kernel[channel:channel + len(kernel0)] = kernel0[:]
+        kernel = kernel[n_side:-n_side]
+        assert len(kernel) == n_channels
+        positive = kernel > 0
+        negative = kernel < 0
+        kernel[positive] /= sum(kernel[positive])
+        kernel[negative] /= -1 * sum(kernel[negative])
+        return kernel
+
+
+def _gaussian0(x, mean, sigma):
+    """Gaussian function."""
+    z = (x - mean) / sigma
+    return np.exp(-z**2 / 2.)
+
+
+def _gaussian1(x, mean, sigma):
+    """First derivative of a gaussian."""
+    z = (x - mean)
+    return -1 * z * _gaussian0(x, mean, sigma)
+
+
+class GaussianPeakFilter(PeakFilter):
+    """A spectral kernel that is the second derivative of a Gaussian.
+
+    g0(x, u, s) = 1 / (sqrt(2 pi) s) * exp(-(x-u)^2/(2 s^2))
+    g1(x, u, s) = -(x - u) / (s^2) * g0
+    g2(x, u, s) = (x - u)^2 / (s^4) * g0 - (1/s^2) * g0
+                = g0(x,u,s) ((x-u)^2/s^2 - 1) / s^2
+
+    integral(-g2, x=x0..x1)
+        = (-g1) for x=x0..x1
+        = g1(x0) - g1(x1)
+
+    """
+
+    def kernel(self, channel, n_channels):
+        """Generate the kernel for the given channel."""
+        fwhm = self.fwhm(channel)
+        sigma = fwhm / 2.355
+        edges = np.arange(n_channels + 1)
+        g1_x0 = _gaussian1(edges[:-1], channel + 0.5, sigma)
+        g1_x1 = _gaussian1(edges[1:], channel + 0.5, sigma)
+        kernel = g1_x0 - g1_x1
+        return kernel
+
+
+class PeakFinder(object):
+    """Find peaks in a spectrum after convolving it with a kernel."""
+
+    def __init__(self, spectrum, kernel, min_sep=5):
+        """Initialize with a spectrum and kernel."""
+        assert min_sep > 0
+        self.min_sep = min_sep
+        self.calculate(spectrum, kernel)
+
+    def reset(self):
+        """Restore peak finder to pristine starting condition."""
+        self.channels = []
+        self.snrs = []
+
+    def calculate(self, spectrum, kernel):
+        """Calculate the convolution of the spectrum with the kernel."""
+        assert isinstance(spectrum, Spectrum)
+        assert isinstance(kernel, PeakFilter)
+        self.spectrum = spectrum
+        self.kernel = kernel
+        self.snr = np.zeros(len(self.spectrum))
+        # calculate the convolution
+        peak_plus_bkg, bkg, signal, noise, snr = \
+            self.kernel.convolve(self.spectrum.counts_vals)
+        self._peak_plus_bkg = peak_plus_bkg
+        self._bkg = bkg
+        self._signal = signal
+        self._noise = noise
+        self.snr = snr.clip(0)
+        self.min_snr = 3.0
+        self.max_num = 20
+        self.min_chan = 0
+        self.max_chan = np.inf
+        self.reset()
+
+    def add_peak(self, chan, snr):
+        """Add a peak at the channelto the list if it is not already there."""
+        new_channel = True
+        for chan2 in self.channels:
+            if abs(chan - chan2) <= self.min_sep:
+                new_channel = False
+        if new_channel:
+            self.channels.append(chan)
+            self.snrs.append(snr)
+        # sort the peaks by channel
+        self.channels = np.array(self.channels)
+        self.snrs = np.array(self.snrs)
+        i = np.argsort(self.channels)
+        self.channels = list(self.channels[i])
+        self.snrs = list(self.snrs[i])
+
+    def plot(self, facecolor='red', linecolor=None, alpha=0.5, peaks=True):
+        """Plot the peak signal-to-noise ratios calculated using the kernel."""
+        if facecolor is not None:
+            plt.fill_between(
+                self.spectrum.channels, self.snr, 0,
+                color=facecolor, alpha=alpha)
+        if linecolor is not None:
+            plt.plot(self.spectrum.channels, self.snr, '-', color=linecolor)
+        if peaks:
+            for chan, snr in zip(self.channels, self.snrs):
+                plt.plot([chan] * 2, [0, snr], 'b-', lw=1.5)
+                plt.plot(chan, snr, 'bo')
+        plt.xlim(0, len(self.spectrum))
+        plt.ylim(0)
+        plt.xlabel('Channels')
+        plt.ylabel('SNR')
+
+    def find_peak(self, channel, frac_range=[0.8, 1.2], min_snr=None):
+        """Find the highest SNR peak within f0*channel and f1*channel."""
+        assert 0 <= channel <= self.spectrum.channels.max()
+        assert 0 < frac_range[0] < 1
+        assert 1 < frac_range[1]
+        assert frac_range[0] < frac_range[1]
+        if min_snr is not None:
+            self.min_snr = min_snr
+        assert self.min_snr > 0.
+        assert self.min_snr <= self.snr.max()
+        chan_range = \
+            (frac_range[0] * channel <= self.spectrum.channels) & \
+            (self.spectrum.channels <= frac_range[1] * channel)
+        peak_snr = self.snr[chan_range].max()
+        if peak_snr < self.min_snr:
+            return None
+        peak_chan = np.where((self.snr == peak_snr) & chan_range)[0][0]
+        self.add_peak(peak_chan, peak_snr)
+        return peak_chan, peak_snr
+
+    def find_peaks(
+            self, min_chan=None, max_chan=None, min_snr=None, max_num=None):
+        """Find the highest SNR peaks in the data."""
+        if min_chan is not None:
+            self.min_chan = min_chan
+        if max_chan is not None:
+            self.max_chan = max_chan
+        assert 0 <= self.min_chan
+        assert self.min_chan < self.max_chan
+        if min_snr is not None:
+            self.min_snr = min_snr
+        assert self.min_snr > 0.
+        assert self.min_snr <= self.snr.max()
+        if max_num is not None:
+            self.max_num = max_num
+        assert self.max_num >= 2
+        # find regions above min_snr
+        if self.snr.max() < self.min_snr:
+            raise Exception(
+                'SNR threshold is {:.3f} but maximum SNR is {:.3f}'.format(
+                    self.min_snr, self.snr.max()))
+        if self.snr[0] >= self.min_snr:
+            self.snr[0] = 0.
+        if self.snr[-1] >= self.min_snr:
+            self.snr[-1] = 0.
+        left_edges = np.where(
+            (self.snr[:-1] < self.min_snr) &
+            (self.snr[1:] >= self.min_snr))[0] + 1
+        right_edges = np.where(
+            (self.snr[:-1] >= self.min_snr) &
+            (self.snr[1:] < self.min_snr))[0]
+        assert len(left_edges) == len(right_edges)
+        # find maximum within each region
+        for left, right in zip(left_edges, right_edges):
+            in_region = (left <= self.spectrum.channels) & \
+                (self.spectrum.channels <= right)
+            peak_snr = self.snr[in_region].max()
+            peak_chan = self.spectrum.channels[
+                (self.snr == peak_snr) & in_region][0]
+            if self.min_chan <= peak_chan <= self.max_chan:
+                self.add_peak(peak_chan, peak_snr)
+        self.channels = np.array(self.channels)
+        self.snrs = np.array(self.snrs)
+        # reduce number of channels to a maximum number max_n
+        i = np.argsort(self.snrs)
+        self.channels = self.channels[i][::-1]
+        self.snrs = self.snrs[i][::-1]
+        self.channels = self.channels[-self.max_num:]
+        self.snrs = self.snrs[-self.max_num:]
+        # sort by channel
+        i = np.argsort(self.channels)
+        self.channels = list(self.channels[i])
+        self.snrs = list(self.snrs[i])
